+++ extension/partiql-extension-ion/src/encode.rs
@@ -4,7 +4,6 @@ use ion_rs::{IonError, IonType, IonWriter};
 use ordered_float::OrderedFloat;
 use partiql_value::{Bag, DateTime, List, Tuple, Value};
 use rust_decimal::Decimal;
-use std::io::Write;
 
 use crate::common::{
     BAG_ANNOT, DATE_ANNOT, MISSING_ANNOT, TIME_ANNOT, TIME_PART_HOUR_KEY, TIME_PART_MINUTE_KEY,
@@ -77,7 +76,7 @@ impl IonEncoderBuilder {
         writer: &'a mut I,
     ) -> Result<Box<dyn ValueEncoder<W, I> + 'a>, IonEncodeError>
     where
-        W: Write + 'a,
+        W: 'a,
         I: IonWriter<Output = W> + 'a,
     {
         let encoder = SimpleIonValueEncoder { writer };
@@ -101,7 +100,6 @@ impl Default for IonEncoderBuilder {
 /// An encoder which will write [`Value`]s as Ion stream values.
 pub trait ValueEncoder<W, I>
 where
-    W: Write,
     I: IonWriter<Output = W>,
 {
     /// A reference to the writer used by this encoder.
@@ -113,7 +111,6 @@ where
 
 trait IonValueEncoder<W, I>: ValueEncoder<W, I>
 where
-    W: Write,
     I: IonWriter<Output = W>,
 {
     fn get_writer(&mut self) -> &mut I;
@@ -152,7 +149,7 @@ where
 
 impl<'a, W, I> ValueEncoder<W, I> for SimpleIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     fn writer(&mut self) -> &mut I {
@@ -167,7 +164,7 @@ where
 
 impl<'a, W, I> ValueEncoder<W, I> for PartiqlEncodedIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     fn writer(&mut self) -> &mut I {
@@ -182,7 +179,7 @@ where
 
 struct SimpleIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     pub(crate) writer: &'a mut I,
@@ -190,7 +187,7 @@ where
 
 impl<'a, W, I> IonValueEncoder<W, I> for SimpleIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     fn get_writer(&mut self) -> &mut I {
@@ -280,7 +277,7 @@ where
 #[inline]
 fn encode_list<'a, W, I, V>(encoder: &'a mut impl IonValueEncoder<W, I>, vals: V) -> IonEncodeResult
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W> + 'a,
     V: Iterator<Item = &'a Value> + 'a,
 {
@@ -299,7 +296,7 @@ fn encode_tuple<'a, W, I>(
     val: &'a Tuple,
 ) -> IonEncodeResult
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W> + 'a,
 {
     encoder.writer().step_in(IonType::Struct)?;
@@ -314,7 +311,7 @@ where
 
 struct PartiqlEncodedIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     inner: SimpleIonValueEncoder<'a, W, I>,
@@ -322,7 +319,7 @@ where
 
 impl<'a, W, I> PartiqlEncodedIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     fn write_date(&mut self, date: &Date) -> IonEncodeResult {
@@ -364,7 +361,7 @@ where
 
 impl<'a, W, I> IonValueEncoder<W, I> for PartiqlEncodedIonValueEncoder<'a, W, I>
 where
-    W: Write,
+    W: 'a,
     I: IonWriter<Output = W>,
 {
     fn get_writer(&mut self) -> &mut I {